# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finwiz",
        "description": "Finwiz API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/user-all-categories": {
            "get": {
                "tags": [
                    "User Categories"
                ],
                "summary": "Get user's categories with pivots",
                "operationId": "070b2a606d0e85dc11b361ce27d365fe",
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/user-categories/{id}": {
            "get": {
                "tags": [
                    "User Categories"
                ],
                "summary": "Get a user category with pivots by ID",
                "operationId": "7fa8d01eae25a20d6cf3864585e9f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user category to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Categories"
                ],
                "summary": "Update a user category by ID",
                "operationId": "7ad23ca9f3004111493970d34312eb56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user category to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Updated Category Name"
                                    },
                                    "limitation": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.75
                                    },
                                    "fixed": {
                                        "type": "boolean",
                                        "format": "number",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/merge-user-categories": {
            "post": {
                "tags": [
                    "User Categories"
                ],
                "summary": "Merge user category ",
                "operationId": "ffed28078cfa707951fbc33859b4c42d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Category Name"
                                    },
                                    "limitation": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "user_category_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": "[1, 2, 3]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}